+++ src/main/java/org/apache/aurora/scheduler/preemptor/PreemptionVictimFilter.java
@@ -113,12 +113,20 @@ public interface PreemptionVictimFilter {
           @Override
           public ResourceBag apply(PreemptionVictim victim) {
             ResourceBag bag = victim.getResourceBag();
+
+            if (victim.getConfig().isSetExecutorConfig()) {
+              // Be pessimistic about revocable resource available if config is not available
+              bag.add(executorSettings.getExecutorOverhead(
+                  victim.getConfig().getExecutorConfig().getName()).orElse(EMPTY));
+            }
+
             if (tierManager.getTier(victim.getConfig()).isRevocable()) {
               // Revocable task CPU cannot be used for preemption purposes as it's a compressible
               // resource. We can still use RAM, DISK and PORTS as they are not compressible.
               bag = bag.filter(IS_MESOS_REVOCABLE.negate());
             }
-            return bag.add(executorSettings.getExecutorOverhead());
+
+            return bag;
           }
         };
 
@@ -197,6 +205,11 @@ public interface PreemptionVictimFilter {
         return Optional.absent();
       }
 
+      ResourceBag overhead = pendingTask.isSetExecutorConfig()
+          ? executorSettings.getExecutorOverhead(
+              pendingTask.getExecutorConfig().getName()).orElse(EMPTY)
+          : EMPTY;
+
       ResourceBag totalResource = slackResources;
       for (PreemptionVictim victim : sortedVictims) {
         toPreemptTasks.add(victim);
@@ -205,8 +218,7 @@ public interface PreemptionVictimFilter {
             new UnusedResource(totalResource, attributes.get()),
             new ResourceRequest(
                 pendingTask,
-                ResourceManager.bagFromResources(pendingTask.getResources())
-                    .add(executorSettings.getExecutorOverhead()),
+                ResourceManager.bagFromResources(pendingTask.getResources()).add(overhead),
                 jobState));
 
         if (vetoes.isEmpty()) {