+++ RELEASE-NOTES.md
@@ -17,12 +17,16 @@
   from Mesos. This has affected rendering of some of the existing attributes. Furthermore, it now
   dumps additional offer attributes including [reservations](http://mesos.apache.org/documentation/latest/reservation/)
   and [persistent volumes](http://mesos.apache.org/documentation/latest/persistent-volume/).
-* The scheduler API now accepts both thrift JSON and binary thrift. If a request is sent with a
+- The scheduler API now accepts both thrift JSON and binary thrift. If a request is sent with a
   `Content-Type` header, or a `Content-Type` header of `application/x-thrift` or `application/json`
   or `application/vnd.apache.thrift.json` the request is treated as thrift JSON. If a request is
   sent with a `Content-Type` header of `application/vnd.apache.thrift.binary` the request is treated
   as binary thrift. If the `Accept` header of the request is `application/vnd.apache.thrift.binary`
   then the response will be binary thrift. Any other value for `Accept` will result in thrift JSON.
+- Scheduler is now able to launch jobs using more than one executor at a time. To use this feature
+  the `-custom_executor_config` flag must point to a JSON file which contains at least one valid
+  executor configuration as detailed in the [configuration](http://aurora.apache.org/documentation/latest/operations/configuration/)
+  documentation.
 
 ### Deprecations and removals: